libssc-{sensor,client}: GTask owns GError
If we propagate a GError through a GTask, we don't own the GError
anymore since GTask assumes ownership. The GError will be freed when
handling the GTask, thus we may not put a g_autoptr there. Otherwise
this GError will be freed twice resulting into a segfault. If we
propagate a GError to an async function to retrieve the GError when the
async task finishes, we must use g_autoptr to make sure that the GError
is freed when the GTask completes.