sensor-*: use g_cond to make thread running signal atomic
This was not thread safe and probably caused some race-y behavior. Based
on glib's docs, g_cond was basically made to handle situations like
this, so let's use it.
This might also fix the "is thread running?" while loop spinning
uncontrollably and burning CPU time, but since that problem is hard(?)
to reproduce, only time will tell.